Fix issue 369 - setAppVolume (#392)
* Fix issue 369 - setAppVolume

* Update CHANGELOG.md

Add fix for #369

+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/GoogleMobileAdsPlugin.java
@@ -439,7 +439,7 @@ public class GoogleMobileAdsPlugin implements FlutterPlugin, ActivityAware, Meth
         result.success(null);
         break;
       case "MobileAds#setAppVolume":
-        flutterMobileAds.setAppVolume(call.<Float>argument("volume"));
+        flutterMobileAds.setAppVolume(call.<Double>argument("volume"));
         result.success(null);
         break;
       case "setImmersiveMode":
